+++ src/core/materials/dto/CreateMaterialRequestDTO.rs
@@ -15,7 +15,7 @@ pub struct CreateMaterialRequest {
 
     pub material_type: MaterialType,
 
-    pub estimated_duration_seconds: Option<i32>,
+    pub estimated_duration_seconds: Option<i64>,
 
     pub is_featured: bool,
     pub is_free: bool,
